End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -43,11 +44,12 @@ public class AuctionsController : ControllerBase{
         return _mapper.Map<AuctionDto>(auction);
 
  }
+ [Authorize]
  [HttpPost]
  public async Task<ActionResult<AuctionDto>>CreateAuction(CreateAuctionDto auctionDto){
    var auction = _mapper.Map<Auction>(auctionDto);
-   //TODO: add current user as seller
-   auction.Seller = "test";
+   //add current user as seller
+   auction.Seller = User.Identity.Name;
    _context.Auctions.Add(auction);
       var newAuction  = _mapper.Map<AuctionDto>(auction);
    await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
@@ -58,7 +60,7 @@ public class AuctionsController : ControllerBase{
    }
    return CreatedAtAction(nameof(GetAuctionById),new{newAuction.Id},newAuction);
  }
-
+ [Authorize]
  [HttpPut("{id}")]
 public async Task<ActionResult> UpdateAuction(Guid id, [FromBody]UpdateAuctionDto updateAuctionDto){
    var auction = await _context.Auctions.Include(x=>x.Item)
@@ -66,7 +68,10 @@ public async Task<ActionResult> UpdateAuction(Guid id, [FromBody]UpdateAuctionDt
    if(auction == null){
       return NotFound();      
    }
-   // TODO: check seller == username
+   //check seller == username
+   if(auction.Seller != User.Identity.Name){
+      return Forbid();
+   }
    auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
    auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
    auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
@@ -79,13 +84,17 @@ public async Task<ActionResult> UpdateAuction(Guid id, [FromBody]UpdateAuctionDt
    }
    return BadRequest("Problem saving changes");
 }
+[Authorize]
 [HttpDelete("{id}")]
 public async Task<ActionResult> DeleteAuction(Guid id){
    var auction = await _context.Auctions.FindAsync(id);
    if(auction == null){
       return NotFound();
    }
-   //TODO: check seller == username
+   //check seller == username
+   if (auction.Seller != User.Identity.Name){
+      return Forbid();
+   }
    _context.Auctions.Remove(auction);
    await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
    var result = await _context.SaveChangesAsync()>0;
